+++ packages/netlify-cms-lib-util/src/index.js
@@ -1,6 +1,6 @@
 export APIError from './APIError';
 export Cursor, { CURSOR_COMPATIBILITY_SYMBOL } from './Cursor';
-export EditorialWorkflowError from './EditorialWorkflowError';
+export EditorialWorkflowError, { EDITORIAL_WORKFLOW_ERROR } from './EditorialWorkflowError';
 export localForage from './localForage';
 export { resolvePath, basename, fileExtensionWithSeparator, fileExtension } from './path';
 export { filterPromises, resolvePromiseProperties, then } from './promise';