+++ packages/netlify-cms-widget-markdown/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.10](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-markdown/compare/netlify-cms-widget-markdown@2.0.9...netlify-cms-widget-markdown@2.0.10) (2018-11-29)
+
+
+### Bug Fixes
+
+* **netlify-cms-widget-markdown:** add missing border radius on toolbar ([#1905](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-markdown/issues/1905)) ([3772171](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-markdown/commit/3772171))
+
+
+
+
+
+## [2.0.9](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-markdown/compare/netlify-cms-widget-markdown@2.0.8...netlify-cms-widget-markdown@2.0.9) (2018-11-12)
+
+
+### Bug Fixes
+
+* **editor-component-image:** fix null on empty markdown image alt ([#1778](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-markdown/issues/1778)) ([9b72419](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-markdown/commit/9b72419))
+* **editor-components:** fix default value processing ([#1848](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-markdown/issues/1848)) ([a0cfa1a](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-markdown/commit/a0cfa1a))
+
+
+
+
+
 <a name="2.0.8"></a>
 ## [2.0.8](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-widget-markdown/compare/netlify-cms-widget-markdown@2.0.7...netlify-cms-widget-markdown@2.0.8) (2018-09-06)